examples/pipelines/demo-question-answering/requirements.txt
pathway[all]
python-dotenv==1.0.1
mpmath==1.3.0
litellm>=1.35
Google-generativeai
Sentence-transformers